requirements.txt
beautifulsoup4
requests
python-dotenv
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langgraph

